seedhelper.js file created
